+++ listenbrainz/db/feedback.py
@@ -2,13 +2,9 @@ import sqlalchemy
 from sqlalchemy import text
 
 from listenbrainz import db
-from listenbrainz.db import timescale
-from listenbrainz.db.mapping import fetch_track_metadata_for_items
+from listenbrainz.db.msid_mbid_mapping import fetch_track_metadata_for_items
 from listenbrainz.db.model.feedback import Feedback
-from listenbrainz import messybrainz as msb_db
-from listenbrainz.messybrainz.data import load_recordings_from_msids
 from typing import List
-from listenbrainz.messybrainz.exceptions import NoDataFoundException
 
 
 def insert(feedback: Feedback):
